requirements.txt
mistralai>=0.0.4
PySide6>=6.5.0
cryptography>=40.0.0
pillow>=9.0.0
pyinstaller>=5.8.0 